Adapting code to run in distributed setting
